backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
pytesseract==0.3.10
omegaconf==2.3.0
pytorch-lightning==2.0.8
